refactor: break attestation wizard into sub-components
[#OCD-3964]
