+++ src/app/api/attestations.jsx
@@ -3,7 +3,7 @@ import { useQuery } from 'react-query';
 import { useAxios } from './axios';
 import options from './options';
 
-const useFetchAttestationData = ({ period }) => {
+const useFetchAttestationForm = ({ period }) => {
   const axios = useAxios();
   return useQuery(['attestations'], async () => {
     const response = await axios.get(`/attestations/periods/${period.id}/form`);
@@ -16,4 +16,4 @@ const useFetchAttestationData = ({ period }) => {
 
 /* eslint-disable import/prefer-default-export */
 // remove eslint disable line when new api methods are added
-export { useFetchAttestationData };
+export { useFetchAttestationForm };
