+++ src/app/themes/util.js
@@ -2,6 +2,9 @@ const util = {
   iconSpacing: {
     marginLeft: '4px',
   },
+  fullWidthGridRow: {
+    gridColumn: '1 / -1',
+  },
 };
 
 export default util;
